feat: Add firstOpen functionality to CacheController
This commit adds the `saveFirstOpen` and `getFirstOpen` methods to the `CacheController` mixin. These methods allow for saving and retrieving the `firstOpen` flag in the cache using the `GetStorage` package. The `firstOpen` flag is used to determine if the app is being opened for the first time.

The `saveFirstOpen` method takes a boolean value `firstOpen` and saves it in the cache using the `CacheManagerKey.FIRST_OPEN` key. The `getFirstOpen` method retrieves the `firstOpen` value from the cache using the same key.

This functionality is useful for implementing first-time user experiences or onboarding flows.